pyproject.toml
[project]
name = "storylord"
version = "0.1.0"
description = "AI-powered story generation with pluggable agents"
authors = [
    {name = "Mark Norman", email = "mark@marknorman.com"},
]
dependencies = ["langchain>=1.2.2", "pydantic-settings>=2.12.0", "langchain-anthropic>=1.3.1", "pyyaml>=6.0", "structlog>=25.5.0", "textual>=7.3.0", "textual-dev>=1.8.0", "langgraph>=1.0.6", "pyright>=1.1.408"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[project.entry-points."storylord.architects"]
default = "agents.architect.architect_default:DefaultArchitect"

[project.entry-points."storylord.narrators"]
default = "agents.narrative.narrator_default:DefaultNarrator"

[project.entry-points."storylord.editors"]
default = "agents.editor.editor_default:DefaultEditor"
simile-smasher = "agents.editor.editor_simile_smasher:SimileSmasherEditor"

[project.entry-points."storylord.character_agents"]
default = "agents.character.character_default:DefaultCharacterAgentType"
mbti = "agents.character.character_mbti:MBTICharacterAgentType"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
health = { composite = [
  "echo '\nCheck project for PDM problems\n----------'",
  "pdm fix --dry-run",
  "echo '\nCheck for outdated packages\n----------'",
  "pdm outdated",
  "echo '\nCheck if lock file is up to date\n----------'",
  "pdm lock --check",
  "echo '\nCheck if installed packages match lock file\n----------'",
  "pdm sync --dry-run",
  "echo '\nShow project information\n----------'",
  "pdm info"
]}
fmt = { composite = [
  "echo '\nFix Lint Violations\n----------'",
  "ruff check --fix --show-fixes .",
  "echo '\nFix Format Violations\n----------'",
  "ruff format .",
] }
lint = { composite = [
  "echo '\nLint Check\n----------'",
  "ruff check -e .",
  "echo '\nFormat Check\n----------'",
  "bash -c 'ruff format --check . || true'",
] }
start = "python src/story_lord.py"
test = "pytest tests/ -v"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
